TST/CLN: centralize module check funcs
