+++ pandas/stats/tests/test_moments.py
@@ -17,11 +17,6 @@ from pandas.compat import range, zip, PY3, StringIO
 
 N, K = 100, 10
 
-def _skip_if_no_scipy():
-    try:
-        import scipy.stats
-    except ImportError:
-        raise nose.SkipTest("no scipy.stats")
 
 class TestMoments(tm.TestCase):
 
@@ -68,7 +63,7 @@ class TestMoments(tm.TestCase):
         self._check_moment_func(mom.rolling_mean, np.mean)
 
     def test_cmov_mean(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_mean
         except ImportError:
@@ -86,7 +81,7 @@ class TestMoments(tm.TestCase):
         assert_series_equal(xp, rs)
 
     def test_cmov_window(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_window
         except ImportError:
@@ -104,7 +99,7 @@ class TestMoments(tm.TestCase):
         assert_series_equal(xp, rs)
 
     def test_cmov_window_corner(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_window
         except ImportError:
@@ -128,7 +123,7 @@ class TestMoments(tm.TestCase):
         self.assertEqual(len(rs), 5)
 
     def test_cmov_window_frame(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_window
         except ImportError:
@@ -141,7 +136,7 @@ class TestMoments(tm.TestCase):
         assert_frame_equal(DataFrame(xp), rs)
 
     def test_cmov_window_na_min_periods(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_window
         except ImportError:
@@ -158,7 +153,7 @@ class TestMoments(tm.TestCase):
         assert_series_equal(xp, rs)
 
     def test_cmov_window_regular(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_window
         except ImportError:
@@ -174,7 +169,7 @@ class TestMoments(tm.TestCase):
             assert_series_equal(Series(xp), rs)
 
     def test_cmov_window_special(self):
-        _skip_if_no_scipy()
+        tm._skip_if_no_scipy()
         try:
             from scikits.timeseries.lib import cmov_window
         except ImportError:
