+++ pandas/tests/test_format.py
@@ -86,18 +86,6 @@ def skip_if_np_version_under1p7():
 
         raise nose.SkipTest('numpy >= 1.7 required')
 
-def _skip_if_no_pytz():
-    try:
-        import pytz
-    except ImportError:
-        raise nose.SkipTest("pytz not installed")
-
-def _skip_if_no_dateutil():
-    try:
-        import dateutil
-    except ImportError:
-        raise nose.SkipTest("dateutil not installed")
-
 
 class TestDataFrameFormatting(tm.TestCase):
     _multiprocess_can_split_ = True
@@ -2930,7 +2918,7 @@ class TestStringRepTimestamp(tm.TestCase):
         self.assertEqual(str(ts_nanos_micros), "1970-01-01 00:00:00.000001200")
 
     def test_tz_pytz(self):
-        _skip_if_no_pytz()
+        tm._skip_if_no_pytz()
 
         import pytz
 
@@ -2944,7 +2932,7 @@ class TestStringRepTimestamp(tm.TestCase):
         self.assertEqual(str(dt_datetime_us), str(Timestamp(dt_datetime_us)))
 
     def test_tz_dateutil(self):
-        _skip_if_no_dateutil()
+        tm._skip_if_no_dateutil()
         import dateutil
         utc = dateutil.tz.tzutc()
