+++ pandas/tseries/tests/test_offsets.py
@@ -9,7 +9,7 @@ from nose.tools import assert_raises
 import numpy as np
 
 from pandas.core.datetools import (
-    bday, BDay, cday, CDay, BQuarterEnd, BMonthEnd,
+    bday, BDay, CDay, BQuarterEnd, BMonthEnd,
     CBMonthEnd, CBMonthBegin,
     BYearEnd, MonthEnd, MonthBegin, BYearBegin,
     QuarterBegin, BQuarterBegin, BMonthBegin, DateOffset, Week,
@@ -40,11 +40,6 @@ def test_monthrange():
     for y in range(2000, 2013):
         for m in range(1, 13):
             assert monthrange(y, m) == calendar.monthrange(y, m)
-
-
-def _skip_if_no_cday():
-    if cday is None:
-        raise nose.SkipTest("CustomBusinessDay not available.")
 
 
 ####
@@ -594,7 +589,7 @@ class TestCustomBusinessDay(Base):
         self.d = datetime(2008, 1, 1)
         self.nd = np.datetime64('2008-01-01 00:00:00Z')
 
-        _skip_if_no_cday()
+        tm._skip_if_no_cday()
         self.offset = CDay()
         self.offset2 = CDay(2)
 
@@ -813,7 +808,7 @@ class CustomBusinessMonthBase(object):
     def setUp(self):
         self.d = datetime(2008, 1, 1)
 
-        _skip_if_no_cday()
+        tm._skip_if_no_cday()
         self.offset = self._object()
         self.offset2 = self._object(2)
